Moved getTargetObjectTypeUniqueFieldOutputKeys to new service
