+++ layers/Engine/packages/component-model/src/Schema/FieldQueryInterpreter.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PoP\ComponentModel\Schema;
 
 use PoP\ComponentModel\DirectiveResolvers\DirectiveResolverInterface;
-use PoP\ComponentModel\Exception\SchemaReferenceException;
 use PoP\ComponentModel\Feedback\EngineIterationFeedbackStore;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedback;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
@@ -102,32 +101,6 @@ class FieldQueryInterpreter extends UpstreamFieldQueryInterpreter implements Fie
     final protected function getObjectSerializationManager(): ObjectSerializationManagerInterface
     {
         return $this->objectSerializationManager ??= $this->instanceManager->getInstance(ObjectSerializationManagerInterface::class);
-    }
-
-    /**
-     * If the TypeResolver is of Union type, and we don't have the object
-     * (eg: when printing the configuration), then generate a list of the
-     * unique field outputs for all the target ObjectTypeResolvers.
-     *
-     * If the TypeResolver is an Object type, to respect the same response,
-     * return an array of a single element, with its own unique field output.
-     *
-     * @return array<string,string>
-     */
-    final public function getTargetObjectTypeUniqueFieldOutputKeys(
-        RelationalTypeResolverInterface $relationalTypeResolver,
-        FieldInterface $field,
-    ): array {
-        $uniqueFieldOutputKeys = [];
-        /** @var ObjectTypeResolverInterface[] */
-        $targetObjectTypeResolvers = $relationalTypeResolver instanceof UnionTypeResolverInterface ?
-            $relationalTypeResolver->getTargetObjectTypeResolvers()
-            : [$relationalTypeResolver];
-
-        foreach ($targetObjectTypeResolvers as $targetObjectTypeResolver) {
-            $uniqueFieldOutputKeys[$targetObjectTypeResolver->getTypeName()] = $field->getOutputKey();
-        }
-        return $uniqueFieldOutputKeys;
     }
 
     /**
