+++ layers/Engine/packages/component-model/src/Schema/FieldQueryInterpreterInterface.php
@@ -15,20 +15,6 @@ use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
 interface FieldQueryInterpreterInterface extends UpstreamFieldQueryInterpreterInterface
 {
     /**
-     * If the TypeResolver is of Union type, and we don't have the object
-     * (eg: when printing the configuration), then generate a list of the
-     * unique field outputs for all the target ObjectTypeResolvers.
-     *
-     * If the TypeResolver is an Object type, to respect the same response,
-     * return an array of a single element, with its own unique field output.
-     *
-     * @return array<string,string>
-     */
-    public function getTargetObjectTypeUniqueFieldOutputKeys(
-        RelationalTypeResolverInterface $relationalTypeResolver,
-        FieldInterface $field,
-    ): array;
-    /**
      * Extract field args without using the schema. It is needed to find out which fieldResolver will process a field, where we can't depend on the schema since this one needs to know who the fieldResolver is, creating an infitine loop
      */
     public function extractStaticFieldArguments(string $field, ?array $variables = null): array;
